Split the 3 abnormal API client methods into their own file.
All three of the methods fall outside the current capability of the
normal state RPC client.  Two of them send binary blobs (raw files) so
they send the RPC requests using raw HTTP connections.  The third
streams a running log file over a websocket.

Conceivably both use cases, as well as the use case of receiving a
binary blob, could be accommodated by the normal RPC client by adapting
the jsoncodec package accordingly.  However, that effort is likely
unwarranted and bears a higher risk than leaving the code alone or
addressing the methods in a way that does not impact the normal RPC
code.  Keeping these methods in a separate file should be enough to
indicate that they are abnormal.